Make Java runtime dependency of `libjvm` optional if supported
If the current version of Bazel supports optional toolchains, the Java
runtime dependency of `libjvm` is optional. If no Java runtime is
registered for the current target platform, the resulting binaries will
fall back to looking for a Java runtime on the host.

With this commit, the minimum supported version of Bazel is raised to
5.0.0.